@@ -70,13 +70,12 @@
     "react-csv": "^2.2.2",
     "react-dom": "18.2.0",
     "react-helmet-async": "^2.0.4",
-    "react-intl": "^5.10.15",
+    "react-intl": "^6.4.4",
     "react-is": "18.2.0",
     "react-moment": "^1.1.1",
     "react-number-format": "^4.9.1",
-    "react-router-dom": "5.3.0",
-    "react-router-dom-v5-compat": "^6.3.0",
-    "react-scroll": "^1.8.9",
+    "react-router-dom": "^6.26.0",
+    "react-scroll": "^1.9.0",
     "react-slick": "^0.29.0",
     "react-swipeable": "^7.0.0",
     "react-use-intercom": "^5.4.0",
@@ -89,9 +88,8 @@
     "sha256": "^0.2.0",
     "string-width": "^4.2.3",
     "slick-carousel": "^1.8.1",
-    "styled-components": "5.3.6",
     "tslib": "^2.3.0",
-    "use-query-params": "^2.1.1",
+    "use-query-params": "^2.2.1",
     "yarn": "^1.22.19"
   },
   "devDependencies": {
@@ -142,16 +140,13 @@
     "@types/http-server": "^0.12.1",
     "@types/jest": "28.1.1",
     "@types/node": "18.11.9",
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.11",
+    "@types/react": "18.2.0",
+    "@types/react-dom": "18.2.0",
     "@types/react-is": "17.0.3",
-    "@types/react-router": "^5.1.17",
-    "@types/react-router-dom": "5.3.3",
-    "@types/react-scroll": "^1.8.6",
-    "@types/react-transition-group": "^4.4.5",
+    "@types/react-scroll": "^1.8.10",
+    "@types/react-transition-group": "^4.4.10",
     "@types/sanitize-html": "^2.6.2",
     "@types/set-interval-async": "^1.0.0",
-    "@types/styled-components": "5.1.26",
     "@types/tryghost__content-api": "^1.3.11",
     "@types/youtube-player": "^5.5.5",
     "@typescript-eslint/eslint-plugin": "5.24.0",
@@ -159,7 +154,6 @@
     "assert": "^2.0.0",
     "babel-jest": "28.1.1",
     "babel-plugin-formatjs": "^10.3.29",
-    "babel-plugin-styled-components": "1.10.7",
     "chai": "^4.3.6",
     "cors": "^2.8.5",
     "crypto-browserify": "^3.12.0",
@@ -205,19 +199,18 @@
     "ts-jest": "28.0.5",
     "ts-node": "10.9.1",
     "typechain": "^8.3.2",
-    "typescript": "4.8.4",
+    "typescript": "^5.5.4",
     "url": "^0.11.0",
     "url-loader": "^4.1.1",
     "webpack": "^5.75.0",
     "webpack-merge": "^5.8.0",
     "wrangler": "^3.61.0"
   },
+  "resolutions": {
+    "@types/react": "18.2.0"
+  },
   "workspaces": [
     "packages/**",
     "apps/**"
-  ],
-  "resolutions": {
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.11"
-  }
+  ]
 }
