@@ -19,7 +19,7 @@ import {
 import { Network } from '@notional-finance/util';
 import { useCallback, useEffect, useState } from 'react';
 import { FormattedMessage } from 'react-intl';
-import { useHistory } from 'react-router';
+import { useNavigate } from 'react-router-dom';
 import { findSortingNum } from './use-all-accounts';
 
 interface VaultAccountData {
@@ -50,7 +50,7 @@ export const useAllVaultAccounts = (selectedNetwork: Network) => {
   >(undefined);
   const { updateNotional } = useNotionalContext();
   const baseCurrency = useFiat();
-  const history = useHistory();
+  const navigate = useNavigate();
   const [healthFactorOptions, setHealthFactorOptions] = useState([]);
   const [vaultNameOptions, setVaultNameOptions] = useState([]);
 
@@ -75,7 +75,6 @@ export const useAllVaultAccounts = (selectedNetwork: Network) => {
   const addressClick = useCallback(
     (address: string, network) => {
       updateNotional({
-        hasSelectedChainError: false,
         wallet: {
           signer: undefined,
           selectedAddress: address,
@@ -85,9 +84,9 @@ export const useAllVaultAccounts = (selectedNetwork: Network) => {
         selectedNetwork: network,
       });
 
-      history.push(`/portfolio/${network}/overview`);
+      navigate(`/portfolio/${network}/overview`);
     },
-    [history, updateNotional]
+    [navigate, updateNotional]
   );
 
   const tableColumns = [
