@@ -1,8 +1,5 @@
 import spindl from '@spindl-xyz/attribution';
-import {
-  useSanctionsBlock,
-  useWalletConnectedNetwork,
-} from '@notional-finance/notionable-hooks';
+import { useWalletConnectedNetwork } from '@notional-finance/notionable-hooks';
 import { IntercomProvider } from 'react-use-intercom';
 import {
   LeveragedYieldDashboard,
@@ -15,14 +12,12 @@ import {
   LiquidityLeveragedDashboard,
 } from '@notional-finance/shared-web';
 import { Web3OnboardProvider } from '@web3-onboard/react';
-import { Redirect, Route, Switch, useParams } from 'react-router';
-import { CompatRouter } from 'react-router-dom-v5-compat';
+import { Routes, Route, Navigate, useParams } from 'react-router-dom';
 import { ServerError } from '../ServerError/server-error';
 import RouteContainer from './RouteContainer';
 import AppLayoutRoute from './AppLayoutRoute';
 import LandingLayoutRoute from './LandingLayoutRoute';
 import { OnboardContext } from '@notional-finance/wallet';
-import { useConnect } from '@notional-finance/wallet/hooks';
 // Feature shell views
 import { AboutUsView } from '@notional-finance/about-us-feature-shell';
 import { LendFixed, LendVariable } from '@notional-finance/lend-feature-shell';
@@ -57,214 +52,398 @@ const RedirectToDefaultNetwork = () => {
   const walletNetwork = useWalletConnectedNetwork();
   const { basePath } = useParams<{ basePath: string }>();
   return (
-    <Redirect
-      path={basePath}
-      from={basePath}
+    <Navigate
       to={`${basePath}/${
         walletNetwork || getDefaultNetworkFromHostname(window.location.hostname)
       }`}
+      replace
     />
   );
 };
 
 const AllRoutes = () => {
-  useSanctionsBlock();
-  // Have this hook here to ensure that all children routes will see updates if the onboard
-  // context changes (there is a useEffect hook inside here listening for changes in the
-  // onboard context)
-  useConnect();
-
   return (
-    <CompatRouter>
-      <RouteContainer>
-        <Switch>
-          <AppLayoutRoute
-            path="/borrow-fixed/:selectedNetwork/:selectedDepositToken"
-            component={BorrowFixed}
-            routeType="Transaction"
-          />
-          <AppLayoutRoute
-            path="/borrow-fixed/:selectedNetwork"
-            component={BorrowFixedDashboard}
-            routeType="Card"
-          />
-          <AppLayoutRoute
-            path="/borrow-variable/:selectedNetwork/:selectedDepositToken"
-            component={BorrowVariable}
-            routeType="Transaction"
-          />
-          <AppLayoutRoute
-            path="/borrow-variable/:selectedNetwork"
-            component={BorrowVariableDashboard}
-            routeType="Card"
-          />
-          <AppLayoutRoute
-            path="/lend-fixed/:selectedNetwork/:selectedDepositToken"
-            component={LendFixed}
-            routeType="Transaction"
-          />
-          <AppLayoutRoute
-            path="/lend-fixed/:selectedNetwork"
-            component={LendFixedDashboard}
-            routeType="Card"
-          />
-          <AppLayoutRoute
-            path="/lend-variable/:selectedNetwork/:selectedDepositToken"
-            component={LendVariable}
-            routeType="Transaction"
-          />
-          <AppLayoutRoute
-            path="/lend-variable/:selectedNetwork"
-            component={LendVariableDashboard}
-            routeType="Card"
-          />
-          <AppLayoutRoute
-            path="/liquidity-variable/:selectedNetwork/:selectedDepositToken"
-            component={LiquidityVariable}
-            routeType="Transaction"
-          />
-          <AppLayoutRoute
-            path="/liquidity-variable/:selectedNetwork"
-            component={LiquidityVariableDashboard}
-            routeType="Card"
-          />
-          <AppLayoutRoute
-            path="/liquidity-leveraged/:selectedNetwork/:action/:selectedDepositToken/:selectedToken"
-            component={LiquidityLeveraged}
-            routeType="Transaction"
-          />
-          <AppLayoutRoute
-            path="/liquidity-leveraged/:selectedNetwork/:action/:selectedDepositToken"
-            component={LiquidityLeveraged}
-            routeType="Transaction"
-          />
-          <AppLayoutRoute
-            path="/liquidity-leveraged/:selectedNetwork"
-            component={LiquidityLeveragedDashboard}
-            routeType="Card"
-          />
-          <AppLayoutRoute
-            path="/vaults/:selectedNetwork/:vaultAddress/:action/:selectedToken"
-            component={VaultView}
-            routeType="Transaction"
-          />
-          <AppLayoutRoute
-            path="/vaults/:selectedNetwork/:vaultAddress/:action"
-            component={VaultView}
-            routeType="Transaction"
-          />
-          <AppLayoutRoute
-            path="/vaults/:selectedNetwork/:vaultAddress"
-            component={VaultView}
-            routeType="Transaction"
-          />
-          <AppLayoutRoute
-            path="/leveraged-yield-farming/:selectedNetwork"
-            component={LeveragedYieldDashboard}
-            routeType="Card"
-          />
-          <AppLayoutRoute
-            path="/leveraged-points-farming/:selectedNetwork"
-            component={LeveragedPointsDashboard}
-            routeType="Card"
-          />
-          <AppLayoutRoute
-            path={`/portfolio/:selectedNetwork/:category/:sideDrawerKey/:selectedToken`}
-            component={PortfolioFeatureShell}
-            routeType="PortfolioTransaction"
-          />
-          <AppLayoutRoute
-            path={`/portfolio/:selectedNetwork/:category/:sideDrawerKey/:selectedToken/:action`}
-            component={PortfolioFeatureShell}
-            routeType="PortfolioTransaction"
-          />
-          <AppLayoutRoute
-            path={`/portfolio/:selectedNetwork/:category/:sideDrawerKey/:selectedToken/:action/:selectedCollateralToken`}
-            component={PortfolioFeatureShell}
-            routeType="PortfolioTransaction"
-          />
-          <AppLayoutRoute
-            path="/portfolio/:selectedNetwork/:category/:sideDrawerKey"
-            component={PortfolioFeatureShell}
-            routeType="PortfolioTransaction"
-          />
-          <AppLayoutRoute
-            path="/portfolio/:selectedNetwork/:category/"
-            component={PortfolioFeatureShell}
-            routeType="Portfolio"
-          />
-          <AppLayoutRoute
-            path="/portfolio/:selectedNetwork"
-            component={PortfolioFeatureShell}
-            routeType="Portfolio"
-          />
-          <AppLayoutRoute
-            path="/markets"
-            component={Markets}
-            routeType="Analytics"
-          />
-          <AppLayoutRoute
-            path="/stake/:selectedDepositToken"
-            component={StakeNOTE}
-            routeType="Note"
-          />
-          <AppLayoutRoute path="/note" component={NoteView} routeType="Note" />
-          <AppLayoutRoute
-            path="/analytics/:category"
-            component={AnalyticsViews}
-            routeType="Analytics"
-          />
-          <AppLayoutRoute
-            path="/analytics"
-            component={AnalyticsViews}
-            routeType="Analytics"
-          />
-          <AppLayoutRoute
-            path="/error"
-            component={ServerError}
-            routeType="Error"
-          />
-          <AppLayoutRoute
-            path="/points-dashboard/:selectedNetwork"
-            component={PointsDashboard}
-            routeType="Contest"
-          />
-          <AppLayoutRoute
-            path="/contest/:selectedNetwork"
-            component={ContestHome}
-            routeType="Contest"
-          />
-          <AppLayoutRoute
-            path="/contest-rules/:selectedNetwork"
-            component={ContestRules}
-            routeType="Contest"
-          />
-          <AppLayoutRoute
-            path="/contest-leaderboard/:selectedNetwork"
-            component={ContestLeaderBoard}
-            routeType="Contest"
-          />
-          <AppLayoutRoute
-            path="/contest-sign-up/:selectedNetwork/:step/"
-            component={ContestSignUp}
-            routeType="Contest"
-          />
-          <AppLayoutRoute
-            path="/contest-sign-up/:selectedNetwork"
-            component={ContestSignUp}
-            routeType="Contest"
-          />
-          <LandingLayoutRoute path="/terms" component={TermsView} />
-          <LandingLayoutRoute path="/privacy" component={PrivacyView} />
-          <LandingLayoutRoute path="/about" component={AboutUsView} />
-          {/* Catches all the card pages that should be redirected to the default network */}
-          <Route path="/:basePath">
-            <RedirectToDefaultNetwork />
-          </Route>
-          <LandingLayoutRoute path="/" component={LandingPageView} />
-        </Switch>
-      </RouteContainer>
-    </CompatRouter>
+    <RouteContainer>
+      <Routes>
+        <Route
+          path="/borrow-fixed/:selectedNetwork/:selectedDepositToken"
+          element={
+            <AppLayoutRoute
+              path="/borrow-fixed/:selectedNetwork/:selectedDepositToken"
+              component={BorrowFixed}
+              routeType="Transaction"
+            />
+          }
+        />
+        <Route
+          path="/borrow-fixed/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/borrow-fixed/:selectedNetwork"
+              component={BorrowFixedDashboard}
+              routeType="Card"
+            />
+          }
+        />
+        <Route
+          path="/borrow-variable/:selectedNetwork/:selectedDepositToken"
+          element={
+            <AppLayoutRoute
+              path="/borrow-variable/:selectedNetwork/:selectedDepositToken"
+              component={BorrowVariable}
+              routeType="Transaction"
+            />
+          }
+        />
+        <Route
+          path="/borrow-variable/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/borrow-variable/:selectedNetwork"
+              component={BorrowVariableDashboard}
+              routeType="Card"
+            />
+          }
+        />
+        <Route
+          path="/lend-fixed/:selectedNetwork/:selectedDepositToken"
+          element={
+            <AppLayoutRoute
+              path="/lend-fixed/:selectedNetwork/:selectedDepositToken"
+              component={LendFixed}
+              routeType="Transaction"
+            />
+          }
+        />
+        <Route
+          path="/lend-fixed/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/lend-fixed/:selectedNetwork"
+              component={LendFixedDashboard}
+              routeType="Card"
+            />
+          }
+        />
+        <Route
+          path="/lend-variable/:selectedNetwork/:selectedDepositToken"
+          element={
+            <AppLayoutRoute
+              path="/lend-variable/:selectedNetwork/:selectedDepositToken"
+              component={LendVariable}
+              routeType="Transaction"
+            />
+          }
+        />
+        <Route
+          path="/lend-variable/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/lend-variable/:selectedNetwork"
+              component={LendVariableDashboard}
+              routeType="Card"
+            />
+          }
+        />
+        <Route
+          path="/liquidity-variable/:selectedNetwork/:selectedDepositToken"
+          element={
+            <AppLayoutRoute
+              path="/liquidity-variable/:selectedNetwork/:selectedDepositToken"
+              component={LiquidityVariable}
+              routeType="Transaction"
+            />
+          }
+        />
+        <Route
+          path="/liquidity-variable/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/liquidity-variable/:selectedNetwork"
+              component={LiquidityVariableDashboard}
+              routeType="Card"
+            />
+          }
+        />
+        <Route
+          path="/liquidity-leveraged/:selectedNetwork/:action/:selectedDepositToken/:selectedToken"
+          element={
+            <AppLayoutRoute
+              path="/liquidity-leveraged/:selectedNetwork/:action/:selectedDepositToken/:selectedToken"
+              component={LiquidityLeveraged}
+              routeType="Transaction"
+            />
+          }
+        />
+        <Route
+          path="/liquidity-leveraged/:selectedNetwork/:action/:selectedDepositToken"
+          element={
+            <AppLayoutRoute
+              path="/liquidity-leveraged/:selectedNetwork/:action/:selectedDepositToken"
+              component={LiquidityLeveraged}
+              routeType="Transaction"
+            />
+          }
+        />
+        <Route
+          path="/liquidity-leveraged/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/liquidity-leveraged/:selectedNetwork"
+              component={LiquidityLeveragedDashboard}
+              routeType="Card"
+            />
+          }
+        />
+        <Route
+          path="/vaults/:selectedNetwork/:vaultAddress/:action/:selectedToken"
+          element={
+            <AppLayoutRoute
+              path="/vaults/:selectedNetwork/:vaultAddress/:action/:selectedToken"
+              component={VaultView}
+              routeType="Transaction"
+            />
+          }
+        />
+        <Route
+          path="/vaults/:selectedNetwork/:vaultAddress/:action"
+          element={
+            <AppLayoutRoute
+              path="/vaults/:selectedNetwork/:vaultAddress/:action"
+              component={VaultView}
+              routeType="Transaction"
+            />
+          }
+        />
+        <Route
+          path="/vaults/:selectedNetwork/:vaultAddress"
+          element={
+            <AppLayoutRoute
+              path="/vaults/:selectedNetwork/:vaultAddress"
+              component={VaultView}
+              routeType="Transaction"
+            />
+          }
+        />
+        <Route
+          path="/leveraged-yield-farming/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/leveraged-yield-farming/:selectedNetwork"
+              component={LeveragedYieldDashboard}
+              routeType="Card"
+            />
+          }
+        />
+        <Route
+          path="/leveraged-points-farming/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/leveraged-points-farming/:selectedNetwork"
+              component={LeveragedPointsDashboard}
+              routeType="Card"
+            />
+          }
+        />
+        <Route
+          path={`/portfolio/:selectedNetwork/:category/:sideDrawerKey/:selectedToken`}
+          element={
+            <AppLayoutRoute
+              path={`/portfolio/:selectedNetwork/:category/:sideDrawerKey/:selectedToken`}
+              component={PortfolioFeatureShell}
+              routeType="PortfolioTransaction"
+            />
+          }
+        />
+        <Route
+          path={`/portfolio/:selectedNetwork/:category/:sideDrawerKey/:selectedToken/:action`}
+          element={
+            <AppLayoutRoute
+              path={`/portfolio/:selectedNetwork/:category/:sideDrawerKey/:selectedToken/:action`}
+              component={PortfolioFeatureShell}
+              routeType="PortfolioTransaction"
+            />
+          }
+        />
+        <Route
+          path={`/portfolio/:selectedNetwork/:category/:sideDrawerKey/:selectedToken/:action/:selectedCollateralToken`}
+          element={
+            <AppLayoutRoute
+              path={`/portfolio/:selectedNetwork/:category/:sideDrawerKey/:selectedToken/:action/:selectedCollateralToken`}
+              component={PortfolioFeatureShell}
+              routeType="PortfolioTransaction"
+            />
+          }
+        />
+        <Route
+          path="/portfolio/:selectedNetwork/:category/:sideDrawerKey"
+          element={
+            <AppLayoutRoute
+              path="/portfolio/:selectedNetwork/:category/:sideDrawerKey"
+              component={PortfolioFeatureShell}
+              routeType="PortfolioTransaction"
+            />
+          }
+        />
+        <Route
+          path="/portfolio/:selectedNetwork/:category/"
+          element={
+            <AppLayoutRoute
+              path="/portfolio/:selectedNetwork/:category/"
+              component={PortfolioFeatureShell}
+              routeType="Portfolio"
+            />
+          }
+        />
+        <Route
+          path="/portfolio/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/portfolio/:selectedNetwork"
+              component={PortfolioFeatureShell}
+              routeType="Portfolio"
+            />
+          }
+        />
+        <Route
+          path="/markets"
+          element={
+            <AppLayoutRoute
+              path="/markets"
+              component={Markets}
+              routeType="Analytics"
+            />
+          }
+        />
+        <Route
+          path="/stake/:selectedDepositToken"
+          element={
+            <AppLayoutRoute
+              path="/stake/:selectedDepositToken"
+              component={StakeNOTE}
+              routeType="Note"
+            />
+          }
+        />
+        <Route
+          path="/note"
+          element={
+            <AppLayoutRoute
+              path="/note"
+              component={NoteView}
+              routeType="Note"
+            />
+          }
+        />
+        <Route
+          path="/analytics/:category"
+          element={
+            <AppLayoutRoute
+              path="/analytics/:category"
+              component={AnalyticsViews}
+              routeType="Analytics"
+            />
+          }
+        />
+        <Route
+          path="/analytics"
+          element={
+            <AppLayoutRoute
+              path="/analytics"
+              component={AnalyticsViews}
+              routeType="Analytics"
+            />
+          }
+        />
+        <Route
+          path="/error"
+          element={
+            <AppLayoutRoute
+              path="/error"
+              component={ServerError}
+              routeType="Error"
+            />
+          }
+        />
+        <Route
+          path="/points-dashboard/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/points-dashboard/:selectedNetwork"
+              component={PointsDashboard}
+              routeType="Contest"
+            />
+          }
+        />
+        <Route
+          path="/contest/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/contest/:selectedNetwork"
+              component={ContestHome}
+              routeType="Contest"
+            />
+          }
+        />
+        <Route
+          path="/contest-rules/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/contest-rules/:selectedNetwork"
+              component={ContestRules}
+              routeType="Contest"
+            />
+          }
+        />
+        <Route
+          path="/contest-leaderboard/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/contest-leaderboard/:selectedNetwork"
+              component={ContestLeaderBoard}
+              routeType="Contest"
+            />
+          }
+        />
+        <Route
+          path="/contest-sign-up/:selectedNetwork/:step/"
+          element={
+            <AppLayoutRoute
+              path="/contest-sign-up/:selectedNetwork/:step/"
+              component={ContestSignUp}
+              routeType="Contest"
+            />
+          }
+        />
+        <Route
+          path="/contest-sign-up/:selectedNetwork"
+          element={
+            <AppLayoutRoute
+              path="/contest-sign-up/:selectedNetwork"
+              component={ContestSignUp}
+              routeType="Contest"
+            />
+          }
+        />
+        <Route
+          path="/terms"
+          element={<LandingLayoutRoute component={TermsView} />}
+        />
+        <Route
+          path="/privacy"
+          element={<LandingLayoutRoute component={PrivacyView} />}
+        />
+        <Route
+          path="/about"
+          element={<LandingLayoutRoute component={AboutUsView} />}
+        />
+        <Route path="/:basePath" element={<RedirectToDefaultNetwork />} />
+        <Route
+          path="/"
+          element={<LandingLayoutRoute component={LandingPageView} />}
+        />
+      </Routes>
+    </RouteContainer>
   );
 };
 
