@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { Box } from '@mui/material';
 import { useNotionalError } from '@notional-finance/notionable-hooks';
 
@@ -8,16 +8,16 @@ interface RouteContainerProps {
 }
 
 const RouteContainer = ({ children }: RouteContainerProps) => {
-  const history = useHistory();
+  const navigate = useNavigate();
   const { error } = useNotionalError();
 
   useEffect(() => {
     if (error) {
-      history.push(
+      navigate(
         `/error?code=${error.code || 500}&msgId=${error.msg || 'unknown'}`
       );
     }
-  }, [error, history]);
+  }, [error, navigate]);
 
   return <Box height="100%">{children}</Box>;
 };
