@@ -12,16 +12,19 @@ import {
   PRODUCTS,
 } from '@notional-finance/util';
 import { FormattedMessage, defineMessage } from 'react-intl';
-import { useHistory } from 'react-router';
+import { useNavigate } from 'react-router-dom';
 import { Box, useTheme } from '@mui/material';
 import { LeafIcon, PointsIcon } from '@notional-finance/icons';
 
-export const useFixedRateGrid = (network: Network, product: PRODUCTS) => {
+export const useFixedRateGrid = (
+  network: Network | undefined,
+  product: PRODUCTS
+) => {
   const {
     yields: { fCashLend, fCashBorrow },
   } = useAllMarkets(network);
   const theme = useTheme();
-  const history = useHistory();
+  const navigate = useNavigate();
   const baseCurrency = useFiat();
   const tokenObj = {};
   const isBorrow = product === PRODUCTS.BORROW_FIXED;
@@ -101,7 +104,7 @@ export const useFixedRateGrid = (network: Network, product: PRODUCTS) => {
       tvlNum: y?.liquidity ? y.liquidity.toFiat(baseCurrency).toNumber() : 0,
       apy: y.totalAPY,
       routeCallback: () =>
-        history.push(`/${product}/${network}/${y.underlying.symbol}`),
+        navigate(`/${product}/${network}/${y.underlying.symbol}`),
     };
   });
 
