@@ -1,33 +1,21 @@
-import { CompatRoute } from 'react-router-dom-v5-compat';
 import { Box, CssBaseline, ThemeProvider, styled } from '@mui/material';
 import { colors, useNotionalTheme } from '@notional-finance/styles';
-
-import { useLocation } from 'react-router';
 import { THEME_VARIANTS } from '@notional-finance/util';
+
+interface LandingLayoutRouteProps {
+  component: React.ComponentType<unknown>;
+}
 
 const LandingLayoutRoute = ({
   component: Component,
-  path,
-}: {
-  component: React.ElementType;
-  path: string;
-  landingLayout?: boolean;
-}) => {
-  const location = useLocation();
+}: LandingLayoutRouteProps) => {
   const notionalTheme = useNotionalTheme(THEME_VARIANTS.DARK);
-
   return (
     <ThemeProvider theme={notionalTheme}>
       <CssBaseline />
-      <CompatRoute
-        path={path}
-        key={location.hash}
-        render={(matchProps: Record<string, unknown>) => (
-          <LandingWrapper>
-            <Component {...matchProps} />
-          </LandingWrapper>
-        )}
-      />
+      <LandingWrapper>
+        <Component />
+      </LandingWrapper>
     </ThemeProvider>
   );
 };
