@@ -8,9 +8,9 @@ import { useWalletActive } from '@notional-finance/wallet';
 
 export function LaunchAppButton() {
   const theme = useNotionalTheme(THEME_VARIANTS.LIGHT);
-  const params = useParams<any>();
+  const params = useParams();
   const walletActive = useWalletActive();
-  const network = params?.selectedNetwork || Network.mainnet;
+  const network = params ? params['selectedNetwork'] : Network.mainnet;
 
   return (
     <Button
